Melhorado exercícios do dia 30.03 - 04.04
Exercício 1 = Valores inteiros
Exercício 8 = Apenas 1 Loop para salvar tanto a diagonal principal, tanto a secundária.

+++ 30.03/src/SintaxeJava1.java
@@ -23,7 +23,7 @@ public class SintaxeJava1 {
         for (int i = 0; i < values; i++) {
 
             System.out.println("[DIGITE] Um valor");
-            int value = keyboard.nextInt();
+            float value = keyboard.nextFloat();
             if (value < 0) {
                 negativeValues++;
             }
